# reports/rag_ingestion_auth_review/openapi.json
{
  "components": {
    "schemas": {
...
      "ExportEvalCaseRequest": {
        "properties": {
          "feedback_id": {
            "title": "Feedback Id",
            "type": "integer"
          }
        },
        "required": [
          "feedback_id"
        ],
        "title": "ExportEvalCaseRequest",
        "type": "object"
      },
      "ExportEvalCaseResponse": {
        "properties": {
          "eval_case": {
            "additionalProperties": true,
            "title": "Eval Case",
            "type": "object"
          },
          "feedback_id": {
            "title": "Feedback Id",
            "type": "integer"
          }
        },
        "required": [
          "feedback_id",
          "eval_case"
        ],
        "title": "ExportEvalCaseResponse",
        "type": "object"
      },
      "FeedbackItem": {
        "properties": {
          "answer_source": {
            "title": "Answer Source",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "expected_reply": {
            "title": "Expected Reply",
            "type": "string"
          },
          "exported": {
            "title": "Exported",
            "type": "boolean"
          },
          "failure_stage": {
            "title": "Failure Stage",
            "type": "string"
          },
          "helpful": {
            "title": "Helpful",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "top1_intent": {
            "title": "Top1 Intent",
            "type": "string"
          }
        },
        "required": [
          "id",
          "request_id",
          "query",
          "reply",
          "helpful",
          "reason",
          "expected_reply",
          "top1_intent",
          "latency_ms",
          "answer_source",
          "failure_stage",
          "exported",
          "created_at"
        ],
        "title": "FeedbackItem",
        "type": "object"
      },
      "FeedbackRequest": {
        "properties": {
          "expected_reply": {
            "default": "",
            "title": "Expected Reply",
            "type": "string"
          },
          "helpful": {
            "title": "Helpful",
            "type": "boolean"
          },
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "reply": {
            "minLength": 1,
            "title": "Reply",
            "type": "string"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "trace": {
            "additionalProperties": true,
            "title": "Trace",
            "type": "object"
          }
        },
        "required": [
          "request_id",
          "query",
          "reply",
          "helpful"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
      "FeedbackResponse": {
        "properties": {
          "feedback_id": {
            "title": "Feedback Id",
            "type": "integer"
          },
          "saved": {
            "title": "Saved",
            "type": "boolean"
          }
        },
        "required": [
          "feedback_id",
          "saved"
        ],
        "title": "FeedbackResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderStateRequest": {
        "properties": {
          "delivery_status": {
            "default": "",
            "title": "Delivery Status",
            "type": "string"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "refund_status": {
            "default": "none",
            "title": "Refund Status",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_label": {
            "default": "",
            "title": "Status Label",
            "type": "string"
          },
          "store_name": {
            "default": "",
            "title": "Store Name",
            "type": "string"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "total": {
            "default": 0.0,
            "title": "Total",
            "type": "number"
          },
          "user_id": {
            "default": "demo_user",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "order_id",
          "status"
        ],
        "title": "OrderStateRequest",
        "type": "object"
      },
      "OrderStateResponse": {
        "properties": {
          "delivery_status": {
            "default": "",
            "title": "Delivery Status",
            "type": "string"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "refund_status": {
            "default": "none",
            "title": "Refund Status",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_label": {
            "default": "",
            "title": "Status Label",
            "type": "string"
          },
          "store_name": {
            "default": "",
            "title": "Store Name",
            "type": "string"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "total": {
            "default": 0.0,
            "title": "Total",
            "type": "number"
          },
          "updated_at": {
            "default": "",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "default": "demo_user",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "order_id",
          "status"
        ],
        "title": "OrderStateResponse",
        "type": "object"
      },
...
      "RecentFeedbackResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedbackItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "count",
          "items"
        ],
        "title": "RecentFeedbackResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "身份只来自 Authorization: Bearer <JWT>。X-User-Role / X-Operator-Id 已失效，仅在日志中记录，不参与任何授权判定。",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LLM Customer Service API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/feedback": {
      "post": {
        "operationId": "create_feedback_feedback_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Feedback",
        "tags": [
          "feedback"
        ]
      }
    },
    "/feedback/export-eval-case": {
      "post": {
        "operationId": "export_eval_case_feedback_export_eval_case_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportEvalCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportEvalCaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Eval Case",
        "tags": [
          "feedback"
        ]
      }
    },
    "/feedback/recent": {
      "get": {
        "operationId": "recent_feedback_feedback_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "helpful",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Helpful"
            }
          },
          {
            "in": "query",
            "name": "intent",
            "required": false,
            "schema": {
              "default": "",
              "title": "Intent",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "failure_stage",
            "required": false,
            "schema": {
              "default": "",
              "title": "Failure Stage",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentFeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent Feedback",
        "tags": [
          "feedback"
        ]
      }
    },
...
    "/orders/{order_id}/state": {
      "get": {
        "operationId": "read_order_state_orders__order_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Order State",
        "tags": [
          "orders"
        ]
      },
      "put": {
        "operationId": "save_order_state_orders__order_id__state_put",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Order State",
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}